% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calPvalue.R
\name{calPvalue}
\alias{calPvalue}
\title{Calculating P values of individual clsters}
\usage{
calPvalue(x, clusters, nsim = 1000)
}
\arguments{
\item{x}{an object of class \code{'matrix'}. Each column corresponds to a sample and each row to a variable.}

\item{clusters}{an integer vector of clustering labels.}

\item{nsim}{integer value. the number of simulation. Default is 10.}
}
\value{
the average of P values.
}
\description{
This function is to calculate the average P values of individual clusters.
}
