% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTiplabs.R
\name{compareTiplabs}
\alias{compareTiplabs}
\title{Tip labels comparisons}
\usage{
compareTiplabs(treelist)
}
\arguments{
\item{treelist}{an object of class \code{multiPhylo}.}
}
\value{
an object of class \code{matrix}. \code{TRUE} indicates that two trees are of the same type.because of the identical tip labels.
}
\description{
This function is to compare tip labels of each trees.
}
\examples{
data(GSE45719_268_count)
processed_data <- getPPdata(GSE45719_268_count)
d <- getDlist(x = t(processed_data), mtd = "euclidean")
b <- getBasicPartitions(d, method = "all")

eq <- compareTiplabs(b$partition)
head(eq)
}
