% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBitrees.R
\name{findBitrees}
\alias{findBitrees}
\title{Tress from bipartition data}
\usage{
findBitrees(treelist, format = "all")
}
\arguments{
\item{treelist}{an object of class \code{multiPhylo}.}

\item{format}{a \code{chanracter} vector represents the method of generating the ultimate tree. \code{"dtree"} represents the method based on the distance matrix. \code{"ptree"} represents the method based on maximum parisomy. \code{"ltree"} represents the method based on likelihood. It must be one or combinations of \code{"dtree"}, \code{"ptree"}, \code{"ltree"}. If users want to generate three formats of trees, \code{"all"} could be selected in this function.}
}
\value{
an object of class \code{multiPhylo} which is a list. The elements of this list are phylo trees.
}
\description{
This function is to find the optimal trees from a bipartition data.
}
