% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDim.R
\name{findDim}
\alias{findDim}
\title{Return index with a specific value}
\usage{
findDim(x, object)
}
\arguments{
\item{x}{an object of class \code{matrix}.}

\item{object}{a user-defined value.}
}
\value{
an object of class \code{matrix} of which the 1st column represents the row index and the second represents the column index.
}
\description{
This function is to return the indexes with a user-defined value.
}
