% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasicPartitions.R
\name{getBasicPartitions}
\alias{getBasicPartitions}
\title{Basic Partitions Producted by Hierarchical Clustering}
\usage{
getBasicPartitions(d, method = "complete")
}
\arguments{
\item{d}{a character \code{dist} or \code{list} which stores a number of objects of class \code{dist}.}

\item{method}{a character vector represents the chosen methods to compute. It must be one or combination of \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"}, \code{"centroid"}. If users want to use 8 methods, \code{"all"} could be selected in this function.}
}
\value{
an object of class \code{list}:
\itemize{
\item partition : an object of class \code{multiPhylo} which includes numbers of basic partitions.
\item name : the combinations of specific distances and algorithms.
}
}
\description{
This function creates basic partitions with different combinations of distance measures and hierarchical methods.
}
