% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE45719_268_count}
\alias{GSE45719_268_count}
\title{Gene expression of the tempory dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 22431 rows and 10 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE45719}
}
\usage{
GSE45719_268_count
}
\description{
A matrix contains the single-cell gene expression data from GSE45719.
}
\details{
A dataset containing 10 cells of almost 22,431 genes.
}
\keyword{datasets}
