% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minPkTree.R
\name{minPkTree}
\alias{minPkTree}
\title{Selecting one tree with a specific k from a set of phylo trees}
\usage{
minPkTree(x, tree, min_k = NULL, max_k = NULL, mcores = NULL)
}
\arguments{
\item{x}{an object of class \code{'matrix'}. Each column corresponds to a sample and each row to a variable.}

\item{tree}{an object of class \code{phylo} or \code{multiPhylo}.}

\item{min_k}{integer value. minimum cluster number to evaluate.}

\item{max_k}{integer value. maximum cluster number to evaluate.}

\item{mcores}{the number of CPU cores.}
}
\value{
an object of class \code{Stree} which is a list.
}
\description{
This function is to find the tree with the optimal k identified by the averaged P value. The averaged P value depends on how different the k clusters are. The smaller the averaged P value is, the more separated the clusters are.
}
