% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCellDataset.R
\name{newCellDataSet}
\alias{newCellDataSet}
\title{newCellDataSet from monocle}
\usage{
newCellDataSet(
  cellData,
  phenoData = NULL,
  featureData = NULL,
  lowerDetectionLimit = 0.1,
  expressionFamily = VGAM::negbinomial.size()
)
}
\description{
newCellDataSet from monocle
}
