% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_sc.R
\name{Norm_sc}
\alias{Norm_sc}
\title{Normalization method for scRNA-seq}
\usage{
Norm_sc(
  data,
  norm_method = "CountToCPM",
  gene_length = NULL,
  species = "human"
)
}
\arguments{
\item{data}{an object of class \code{matrix} with cells on columns and genes on rows.}

\item{norm_method}{a character string. This could be one of \code{"CountToCPM"}, \code{"CountToFPKM"}, \code{"CountToCPM"}, \code{"CPMToFPKM"} or \code{"FPKMToCPM"}.}

\item{gene_length}{an numberic vetcor of gene length.}

\item{species}{a character string.It could be one of \code{"huaman"} or \code{"mouse"}}
}
\value{
a normalized matrix.
}
\description{
Normalization method for scRNA-seq
}
