% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDistMethods.R
\name{showDistMethods}
\alias{showDistMethods}
\title{disatnce measures used in scTreeMerge}
\usage{
showDistMethods()
}
\description{
It returns the names of the measures that can be applied to compute distances using the \code{getDlist} function.
}
\examples{
showDistMethods()
}
