% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stree.R
\name{Stree}
\alias{Stree}
\alias{Stree-class}
\title{The Stree class}
\description{
S4 class that holds parameters for the supertree.
}
\section{Parameters}{


The Stree class defines the following parameters:

\describe{
    \item{\code{raw}}{The matrix of P values.}
    \item{\code{tree}}{The structure of the supertree.}
    \item{\code{k}}{The integer value of the optimal cluster number.}
    \item{\code{pvalue}}{The numeric value of individual clsters with the optimal cluster number.}
}
}

