% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTiplabs.R
\name{compareTiplabs}
\alias{compareTiplabs}
\title{Tip labels comparisons}
\usage{
compareTiplabs(treelist)
}
\arguments{
\item{treelist}{an object of class \code{multiPhylo}.}
}
\value{
an object of class \code{matrix}. \code{TRUE} indicates that two trees are of the same type.because of the identical tip labels.
}
\description{
This function is to compare tip labels of each trees.
}
