% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPOPtrees.R
\name{findPOPtrees}
\alias{findPOPtrees}
\title{Find the popular trees}
\usage{
findPOPtrees(treelist)
}
\arguments{
\item{treelist}{an object of class \code{multiPhylo}.}
}
\value{
an object of class \code{list}.
}
\description{
This function is to find the popular trees with the frequency of tree types.
}
