% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDlist.R
\name{getDlist}
\alias{getDlist}
\title{Distance computation with lots of disatnce measures.}
\usage{
getDlist(x, mtd = "euclidean", p = 2)
}
\arguments{
\item{x}{an object of class \code{matrix}. Each column corresponds to a sample and each row to a variable.}

\item{mtd}{a character string that represents the distance measure to be used. It could be one or some of the 47 measures. If users want to use all the measures, \code{all} could be selected in this function. For more information, please see Details.}

\item{p}{The power of Minkowski distance. Note that setting p = 1 is equivalent to calculating the Manhattan distance and setting p = 2 is equivalent to calculating the Euclidean distance. Default: \code{p = 2}.}
}
\value{
an object of class \code{"dist"} or an object if class \code{"list"}.
}
\description{
This is a distance matrix computation function which containins 47 measures. It computes between the rows of a matrix and return the distance matrix computed by choosing the specific measure.
}
\details{
This function implements 47 measures to quantify the disatnce between two objects:
\itemize{
\item Forty-six measures from Philentropy package
\itemize{
\item euclidean
\item manhattan
\item minkowski
\item chebyshev
\item sorensen
\item gower
\item soergel
\item kulczynski_d
\item canberra
\item lorentzian
\item intersection
\item non-intersection
\item wavehedges
\item czekanowski
\item motyka
\item kulczynski_s
\item tanimoto
\item ruzicka
\item inner_product
\item harmonic_mean
\item cosine
\item hassebrook
\item jaccard
\item dice
\item fidelity
\item bhattacharyya
\item hellinger
\item matusita
\item squared_chord
\item squared_euclidean
\item pearson
\item neyman
\item squared_chi
\item prob_symm
\item divergence
\item clark
\item additive_symm
\item kullback-leibler
\item jeffreys
\item k_divergence
\item topsoe
\item jensen-shannon
\item jensen_difference
\item taneja
\item kumar-johnson
\item avg
}

\item One measure from stats package
\itemize{
\item maximum
}
}
}
\examples{
data(GSE45719_268_count)
processed_data <- getPPdata(GSE45719_268_count)

d <- getDlist(x = t(processed_data), mtd = "euclidean")
class(d)
d <- getDlist(x = t(processed_data), mtd = "minkowski", p = 2)
class(d)
# make sure that all vectors sum up to 1.0 ...
d <- getDlist(x = t(processed_data), mtd = "matusita")

d <- getDlist(x = t(processed_data), mtd = c("euclidean", "manhattan"))
class(d)
d[[1]]
class(d[[1]])

d <- getDlist(x = t(processed_data), mtd = "all")
class(d)
}
