% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minPkTree.R
\name{minPkTree}
\alias{minPkTree}
\title{Selecting one tree with a specific k from a set of phylo trees}
\usage{
minPkTree(x, tree, min_k = NULL, max_k = NULL)
}
\arguments{
\item{x}{an object of class \code{'matrix'}. Each column corresponds to a sample and each row to a variable.}

\item{tree}{an object of class \code{phylo} or \code{multiPhylo}.}

\item{min_k}{integer value. minimum cluster number to evaluate.}

\item{max_k}{integer value. maximum cluster number to evaluate.}
}
\value{
an object of class \code{Stree} which is a list.
}
\description{
This function is to find the tree with the optimal k identified by the averaged P value. The averaged P value depends on how different the k clusters are. The smaller the averaged P value is, the more separated the clusters are.
}
\examples{
data(GSE45719_268_count)
processed_data <- getPPdata(GSE45719_268_count)
d <- getDlist(x = t(processed_data), mtd = c("maximum", "euclidean", "manhattan",
                                           "minkowski", "chebyshev", "sorensen",
                                           "gower", "soergel", "kulczynski_d",
                                           "canberra", "lorentzian", "intersection",
                                           "non-intersection", "wavehedges", "czekanowski"))
b <- getBasicPartitions(d, method = "all")
m <- findBitrees(b$partition, "all")
s <- minPkTree(x = processed_data, tree = m, min_k = 2, max_k = 3, iteration = 10)
class(s)
str(s)
}
