% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigclust_n.R
\name{sigclust_n}
\alias{sigclust_n}
\title{Modified sigclust}
\usage{
sigclust_n(x, nsim, label)
}
\arguments{
\item{x}{an object of class \code{matrix}. Each column corresponds to a variable and each row to a sample.}

\item{nsim}{integer value. The number of simulation.}

\item{label}{a numeric, integer vector of 1s and 2s with length \code{nrow(x)} which indicates given cluster labels.}
}
\value{
an object of class \code{list}.
}
\description{
This is a modified function.
}
