% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2Bidata.R
\name{tree2Bidata}
\alias{tree2Bidata}
\title{Extracting tree bipartitions}
\usage{
tree2Bidata(tree)
}
\arguments{
\item{tree}{an object of class \code{"multiPhylo"} or \code{"phylo"}. Note that trees in \code{"multiPhylo"} should have the same tip labels.}
}
\value{
an object of class \code{data.frame}. Its columns represent samples and rows represent different bipartitions.
}
\description{
This function is to extracting the information of bipartitions found in a series of trees.
}
\examples{
data(GSE45719_268_count)
processed_data <- getPPdata(GSE45719_268_count)
d <- getDlist(x = t(processed_data), mtd = c("maximum", "euclidean", "manhattan",
                                           "minkowski", "chebyshev", "sorensen",
                                           "gower", "soergel", "kulczynski_d",
                                           "canberra", "lorentzian", "intersection",
                                           "non-intersection", "wavehedges", "czekanowski"))
b <- getBasicPartitions(d, method = "all")
tree <- b$partition
bidat <- tree2Bidata(tree)
str(bidat)
}
